pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "da3-obsidian"
dynamic = ["version"]
description = "A clean and modular framework developed by the Advanced Analytics (DA3) team at Allianz Partners."
readme = "README.md"
requires-python = ">=3.12"
authors = [
  {name = "Advanced Analytics (DA3) - Allianz Partners", email = "da3-partners@allianz.com"},
]
maintainers = [
  {name = "Abhishek Agarwal", email = "abhishek.agarwal@allianz.com"},
]
keywords = ["text-extraction", "ocr", "pdf", "anonymization", "pii"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core dependencies
    "numpy>=2.2.0",
    "Pillow>=11.2.1",
    
    # Text extraction dependencies
    "pdfplumber>=0.11.7",
    "opencv-python>=4.11.0.86",
    "pytesseract>=0.3.13",
    "deskew>=1.5.3",
    "imutils>=0.5.4",
    
    # Quality checking dependencies
    "nltk>=3.9.1",
    "pyspellchecker>=0.8.3",
    
    # Text anonymization dependencies
    "presidio-analyzer>=2.2.358",
    "presidio-anonymizer>=2.2.358",
    "spacy>=3.8.4",

    # Forensic analysis dependencies
    "pikepdf>=9.9.0",
    "pdf2image>=1.17.0",
    "exifread>=3.3.1",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",

    # Email preprocessing dependencies
    "extract-msg==0.54.1",
    
    # Additional utilities
    "regex>=2024.11.6",

    # Azure Document Intelligence dependencies
    "azure-ai-documentintelligence>=1.0.2",
    "azure-core>=1.34.0",
    "pyyaml>=6.0.2",

    # Saerch Module
    "RapidFuzz>=3.13.0",
    "scikit-learn>=1.7.1",

    "temporalio>=1.5.0",
]

[project.optional-dependencies]
spacy-models = [
    "en_core_web_lg>=3.8.0",
    "fr_core_news_lg>=3.8.0",
    "es_core_news_lg>=3.8.0",
    "it_core_news_lg>=3.8.0",
]
dev = [
    "pytest==8.3.2",
    "pytest-cov==5.0.0",
    "pytest-asyncio==0.23.8",
    "black[d,jupyter]==25.1.0",
    "ruff==0.12.3",
]


[project.urls]
Repository = "https://github.developer.allianz.io/azpa-data-platform/cto-da3-obsidian"
Documentation = "https://github.developer.allianz.io/azpa-data-platform/cto-da3-obsidian"
Issues = "https://github.developer.allianz.io/azpa-data-platform/cto-da3-obsidian/issues"

[tool.hatch.version]
source = "vcs"

[tool.black]
line-length = 120
target-version = ["py312"]
preview = true

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["da3_obsidian"]

[tool.coverage.run]
source = ["src/da3_obsidian"]

[[tool.uv.index]]
name = "nexus-da3"
url = "https://nexus-azp-cto-dataanalytics-shdvp.regional.devops-services.gwc1.azure.aztec.cloud.allianz/repository/pypi-da3/simple"
default = false
explicit = true

[tool.uv.sources]
en_core_web_lg = { index = "nexus-da3" }
fr_core_news_lg = { index = "nexus-da3" }
es_core_news_lg = { index = "nexus-da3" }
it_core_news_lg = { index = "nexus-da3" }
